+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Show whether the user is currently attending with a green chip.
+- `BETA` tag to parking barrier card.
+
+### Changed
+
+- Add a linear gradient and change the period selection button on presence graph.
+
+### Fixed
+
+- Typo in `Déverrouiller` in French.
+
 ## [1.2.0] - 2024-02-23
 
 ### Added